+++ libpolys/coeffs/modulop.cc
@@ -30,7 +30,7 @@ static const n_coeffType ID = n_Zp;
 BOOLEAN npGreaterZero (number k, const coeffs r);
 number  npMult        (number a, number b, const coeffs r);
 number  npInit        (long i, const coeffs r);
-int     npInt         (number &n, const coeffs r);
+long    npInt         (number &n, const coeffs r);
 number  npAdd         (number a, number b,const coeffs r);
 number  npSub         (number a, number b,const coeffs r);
 void    npPower       (number a, int i, number * result,const coeffs r);
@@ -142,12 +142,12 @@ number npInit (long i, const coeffs r)
 /*2
  * convert a number to an int in (-p/2 .. p/2]
  */
-int npInt(number &n, const coeffs r)
+long npInt(number &n, const coeffs r)
 {
   n_Test(n, r);
 
-  if ((long)n > (((long)r->ch) >>1)) return (int)((long)n -((long)r->ch));
-  else                               return (int)((long)n);
+  if ((long)n > (((long)r->ch) >>1)) return ((long)n -((long)r->ch));
+  else                               return ((long)n);
 }
 
 number npAdd (number a, number b, const coeffs r)
